refactor(bot): :building_construction: moved commands to own files
